Move TestComponentViewType to test fixtures
COAND-911
