+++ blik/src/test/java/com/adyen/checkout/blik/BlikComponentTest.kt
@@ -20,7 +20,7 @@ import com.adyen.checkout.components.core.internal.PaymentComponentEvent
 import com.adyen.checkout.test.LoggingExtension
 import com.adyen.checkout.test.TestDispatcherExtension
 import com.adyen.checkout.test.extensions.invokeOnCleared
-import com.adyen.checkout.ui.core.internal.test.TestComponentViewType
+import com.adyen.checkout.ui.core.internal.ui.TestComponentViewType
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.test.runTest
