+++ twint/src/test/java/com/adyen/checkout/twint/TwintActionComponentTest.kt
@@ -12,7 +12,7 @@ import com.adyen.checkout.test.extensions.invokeOnCleared
 import com.adyen.checkout.test.extensions.test
 import com.adyen.checkout.twint.internal.ui.TwintComponentViewType
 import com.adyen.checkout.twint.internal.ui.TwintDelegate
-import com.adyen.checkout.ui.core.internal.test.TestComponentViewType
+import com.adyen.checkout.ui.core.internal.ui.TestComponentViewType
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.test.runTest
