+++ wechatpay/src/test/java/com/adyen/checkout/wechatpay/WeChatPayActionComponentTest.kt
@@ -20,7 +20,7 @@ import com.adyen.checkout.components.core.internal.ActionComponentEventHandler
 import com.adyen.checkout.test.LoggingExtension
 import com.adyen.checkout.test.TestDispatcherExtension
 import com.adyen.checkout.test.extensions.invokeOnCleared
-import com.adyen.checkout.ui.core.internal.test.TestComponentViewType
+import com.adyen.checkout.ui.core.internal.ui.TestComponentViewType
 import com.adyen.checkout.wechatpay.internal.ui.WeChatComponentViewType
 import com.adyen.checkout.wechatpay.internal.ui.WeChatDelegate
 import kotlinx.coroutines.ExperimentalCoroutinesApi
